# Use fast textdocument for caching