This PR improves the performance of the `CFML: Refresh cache for workspace definitions` command that runs on startup. This is the "Caching components" progress meter that appears.

closes cfmleditor/cfmleditor#37

The only change to existing code is to swap `workspace.openTextDocument` with `LSTextDocument.openTextDocument`.

Performance measurements on a code base with 2000+ components:
- Before: 43.9 seconds (with various other extensions enabled)
- Before: 27.1 seconds (best case with all other extensions disabled)
- After: 15.9 seconds (same with or without other extensions)

The problem:

* `workspace.openTextDocument` is heavily tied into the VS Code editor and triggers all the normal events you would get from opening a file. This includes every extension that is listening for files to be opened.
* This is slow and causes a lot of garbage collection
* A listener on another extension can cause the caching step to take orders of magnitude longer. Think 5 minutes instead of 30 seconds.

The solution:

* VS Code has `vscode-languageserver-textdocument`, which is used for this kind of document parsing on language servers that can't access the editor context
* It would be great if we could just use this instead, but it doesn't implement all the methods for the `vscode.TextDocument` interface
* Rewriting the caching code to use `vscode-languageserver-textdocument` is impractical
* Add `LSTextDocument` (LS = Language Server) as a wrapper around `vscode-languageserver-textdocument` that implements the full `vscode.TextDocument` interface

Notes:

* Used https://github.com/microsoft/vscode and https://github.com/microsoft/deoptexplorer-vscode as references for tricky methods (both MIT licensed like CFML Editor)
* `src/utils/wordHelpers.ts` is copied entirely, with the only change being to add a notice at the top
* There are other places in the code that still use `workspace.openTextDocument`, but performance gains from changing them may not be as noticeable